server: add routes and error modules, add run_guardrails_server
