+++ src/server.rs
@@ -14,61 +14,203 @@
  limitations under the License.
 
 */
+use std::{collections::HashSet, net::SocketAddr, path::PathBuf, sync::Arc};
 
-use std::{
-    collections::{HashMap, HashSet},
-    convert::Infallible,
-    error::Error as _,
-    net::SocketAddr,
-    path::PathBuf,
-    sync::Arc,
-};
-
-use axum::{
-    Json, Router,
-    extract::{Query, Request, State, rejection::JsonRejection},
-    http::{HeaderMap, StatusCode},
-    response::{
-        IntoResponse, Response,
-        sse::{Event, KeepAlive, Sse},
-    },
-    routing::{get, post},
-};
-use axum_extra::{extract::WithRejection, json_lines::JsonLines};
-use futures::{
-    Stream, StreamExt,
-    stream::{self, BoxStream},
-};
+use axum::{Router, extract::Request, http::HeaderMap};
 use hyper::body::Incoming;
 use hyper_util::rt::{TokioExecutor, TokioIo};
-use opentelemetry::trace::TraceContextExt;
-use tokio::{net::TcpListener, signal, sync::mpsc};
+use tokio::{net::TcpListener, signal};
 use tokio_rustls::TlsAcceptor;
-use tokio_stream::wrappers::ReceiverStream;
 use tower::Service;
 use tower_http::trace::TraceLayer;
-use tracing::{Span, debug, error, info, warn};
-use tracing_opentelemetry::OpenTelemetrySpanExt;
+use tracing::{debug, error, info, warn};
 
-use crate::{
-    clients::openai::{ChatCompletionsRequest, ChatCompletionsResponse},
-    models::{self, InfoParams, InfoResponse, StreamingContentDetectionRequest},
-    orchestrator::{
-        self, Orchestrator,
-        handlers::{chat_completions_detection::ChatCompletionsDetectionTask, *},
-    },
-    utils,
-};
+use crate::orchestrator::Orchestrator;
 
 mod tls;
 use tls::configure_tls;
+mod errors;
+mod routes;
+pub use errors::Error;
 
-const API_PREFIX: &str = r#"/api/v1/task"#;
-// New orchestrator API
-const TEXT_API_PREFIX: &str = r#"/api/v2/text"#;
+/// Configures and runs orchestrator servers.
+pub async fn run(
+    guardrails_addr: SocketAddr,
+    health_addr: SocketAddr,
+    tls_cert_path: Option<PathBuf>,
+    tls_key_path: Option<PathBuf>,
+    tls_client_ca_cert_path: Option<PathBuf>,
+    orchestrator: Orchestrator,
+) -> Result<(), Error> {
+    let state = Arc::new(ServerState::new(orchestrator));
+    let health_handle = run_health_server(health_addr, state.clone()).await;
+    let guardrails_handle = run_guardrails_server(
+        guardrails_addr,
+        tls_cert_path,
+        tls_key_path,
+        tls_client_ca_cert_path,
+        state.clone(),
+    )
+    .await;
+    // Await server shutdown
+    let _ = tokio::join!(health_handle, guardrails_handle);
+    info!("shutdown complete");
+    Ok(())
+}
 
-const PACKAGE_VERSION: &str = env!("CARGO_PKG_VERSION");
-const PACKAGE_NAME: &str = env!("CARGO_PKG_NAME");
+/// Configures and runs health server.
+async fn run_health_server(
+    addr: SocketAddr,
+    state: Arc<ServerState>,
+) -> tokio::task::JoinHandle<()> {
+    info!("starting health server on {addr}");
+    let app = routes::health_router(state);
+    let listener = TcpListener::bind(&addr)
+        .await
+        .unwrap_or_else(|_| panic!("starting health server failed: could not bind to {addr}"));
+    let server =
+        axum::serve(listener, app.into_make_service()).with_graceful_shutdown(shutdown_signal());
+    tokio::task::spawn(async { server.await.expect("health server crashed!") })
+}
+
+/// Configures and runs guardrails server.
+async fn run_guardrails_server(
+    addr: SocketAddr,
+    tls_cert_path: Option<PathBuf>,
+    tls_key_path: Option<PathBuf>,
+    tls_client_ca_cert_path: Option<PathBuf>,
+    state: Arc<ServerState>,
+) -> tokio::task::JoinHandle<()> {
+    info!("starting guardrails server on {addr}");
+    let router = routes::guardrails_router(state);
+    let app = router.layer(
+        TraceLayer::new_for_http()
+            .make_span_with(crate::utils::trace::incoming_request_span)
+            .on_request(crate::utils::trace::on_incoming_request)
+            .on_response(crate::utils::trace::on_outgoing_response)
+            .on_eos(crate::utils::trace::on_outgoing_eos),
+    );
+    let listener = TcpListener::bind(&addr)
+        .await
+        .unwrap_or_else(|_| panic!("starting guardrails server failed: could not bind to {addr}"));
+    let tls_config = configure_tls(tls_cert_path, tls_key_path, tls_client_ca_cert_path);
+    let shutdown_signal = shutdown_signal();
+    if let Some(tls_config) = tls_config {
+        serve_with_tls(app, listener, tls_config, shutdown_signal)
+    } else {
+        let server =
+            axum::serve(listener, app.into_make_service()).with_graceful_shutdown(shutdown_signal);
+        tokio::task::spawn(async { server.await.expect("guardrails server crashed!") })
+    }
+}
+
+/// Serve the service with the supplied listener, TLS config, and shutdown signal.
+/// Based on https://github.com/tokio-rs/axum/blob/main/examples/low-level-rustls/src/main.rs
+fn serve_with_tls<F>(
+    app: Router,
+    listener: TcpListener,
+    tls_config: Arc<rustls::ServerConfig>,
+    shutdown_signal: F,
+) -> tokio::task::JoinHandle<()>
+where
+    F: Future<Output = ()> + Send + 'static,
+{
+    let tls_acceptor = TlsAcceptor::from(tls_config);
+    tokio::spawn(async move {
+        let graceful = hyper_util::server::graceful::GracefulShutdown::new();
+        let builder = hyper_util::server::conn::auto::Builder::new(TokioExecutor::new());
+        let mut signal = std::pin::pin!(shutdown_signal);
+        loop {
+            let tower_service = app.clone();
+            let tls_acceptor = tls_acceptor.clone();
+            // Wait for new tcp connection
+            let (cnx, addr) = tokio::select! {
+                res = listener.accept() => {
+                    match res {
+                        Ok(res) => res,
+                        Err(err) => {
+                            error!("error accepting tcp connection: {err}");
+                            continue;
+                        }
+                    }
+                }
+                _ = &mut signal => {
+                    debug!("graceful shutdown signal received");
+                    break;
+                }
+            };
+            // Wait for tls handshake
+            let stream = tokio::select! {
+                res = tls_acceptor.accept(cnx) => {
+                    match res {
+                        Ok(stream) => stream,
+                        Err(err) => {
+                            error!("error accepting connection on handshake: {err}");
+                            continue;
+                        }
+                    }
+                }
+                _ = &mut signal => {
+                    debug!("graceful shutdown signal received");
+                    break;
+                }
+            };
+            // `TokioIo` converts between Hyper's own `AsyncRead` and `AsyncWrite` traits
+            let stream = TokioIo::new(stream);
+            let hyper_service = hyper::service::service_fn(move |request: Request<Incoming>| {
+                // Clone necessary since hyper's `Service` uses `&self` whereas
+                // tower's `Service` requires `&mut self`
+                tower_service.clone().call(request)
+            });
+            let conn = builder.serve_connection_with_upgrades(stream, hyper_service);
+            let fut = graceful.watch(conn.into_owned());
+            tokio::spawn(async move {
+                if let Err(err) = fut.await {
+                    warn!("error serving connection from {}: {}", addr, err);
+                }
+            });
+        }
+        tokio::select! {
+            () = graceful.shutdown() => {
+                debug!("graceful shutdown completed");
+            },
+            () = tokio::time::sleep(std::time::Duration::from_secs(10)) => {
+                debug!("graceful shutdown timed out, aborting...");
+            }
+        }
+    })
+}
+
+/// Shutdown signal handler
+async fn shutdown_signal() {
+    let ctrl_c = async {
+        signal::ctrl_c()
+            .await
+            .expect("failed to install Ctrl+C handler");
+    };
+    #[cfg(unix)]
+    let terminate = async {
+        signal::unix::signal(signal::unix::SignalKind::terminate())
+            .expect("failed to install signal handler")
+            .recv()
+            .await;
+    };
+    #[cfg(not(unix))]
+    let terminate = std::future::pending::<()>();
+    tokio::select! {
+        _ = ctrl_c => {},
+        _ = terminate => {},
+    }
+    info!("signal received, starting graceful shutdown");
+}
+
+fn filter_headers(passthrough_headers: &HashSet<String>, headers: HeaderMap) -> HeaderMap {
+    headers
+        .iter()
+        .filter(|(name, _)| passthrough_headers.contains(&name.as_str().to_lowercase()))
+        .map(|(name, value)| (name.clone(), value.clone()))
+        .collect()
+}
 
 /// Server shared state
 pub struct ServerState {
@@ -79,599 +221,4 @@ impl ServerState {
     pub fn new(orchestrator: Orchestrator) -> Self {
         Self { orchestrator }
     }
-}
-
-/// Run the orchestrator server
-pub async fn run(
-    http_addr: SocketAddr,
-    health_http_addr: SocketAddr,
-    tls_cert_path: Option<PathBuf>,
-    tls_key_path: Option<PathBuf>,
-    tls_client_ca_cert_path: Option<PathBuf>,
-    orchestrator: Orchestrator,
-) -> Result<(), Error> {
-    // Overall, the server setup and run does a couple of steps:
-    // (1) Sets up a HTTP server (without TLS) for the health endpoint
-    // (2) Sets up a HTTP(s) server for the main guardrails endpoints
-    //     (2a) Configures TLS or mTLS depending on certs/key provided
-    //     (2b) Adds server routes
-    //     (2c) Generate the server task based on whether or not TLS is configured
-    // (3) Launch each server as a separate task
-    // NOTE: axum::serve is used for servers without TLS since it is designed to be
-    // simple and not allow for much configuration. To allow for TLS configuration
-    // with rustls, the hyper and tower crates [what axum is built on] had to
-    // be used directly
-
-    let state = Arc::new(ServerState::new(orchestrator));
-
-    // (1) Separate HTTP health server without TLS for probes
-    let health_handle = run_health_server(health_http_addr, state.clone()).await;
-
-    // (2) Main guardrails server
-    // (2a) Configure TLS if requested
-    let tls_config = configure_tls(tls_cert_path, tls_key_path, tls_client_ca_cert_path);
-
-    // (2b) Add main guardrails server routes
-    let mut router = Router::new()
-        .route(
-            &format!("{}/classification-with-text-generation", API_PREFIX),
-            post(classification_with_gen),
-        )
-        .route(
-            &format!("{}/detection/stream-content", TEXT_API_PREFIX),
-            post(stream_content_detection),
-        )
-        .route(
-            &format!(
-                "{}/server-streaming-classification-with-text-generation",
-                API_PREFIX
-            ),
-            post(stream_classification_with_gen),
-        )
-        .route(
-            &format!("{}/generation-detection", TEXT_API_PREFIX),
-            post(generation_with_detection),
-        )
-        .route(
-            &format!("{}/detection/content", TEXT_API_PREFIX),
-            post(detection_content),
-        )
-        .route(
-            &format!("{}/detection/chat", TEXT_API_PREFIX),
-            post(detect_chat),
-        )
-        .route(
-            &format!("{}/detection/context", TEXT_API_PREFIX),
-            post(detect_context_documents),
-        )
-        .route(
-            &format!("{}/detection/generated", TEXT_API_PREFIX),
-            post(detect_generated),
-        );
-
-    // If chat generation is configured, enable the chat completions detection endpoint.
-    if state.orchestrator.config().chat_generation.is_some() {
-        info!("Enabling chat completions detection endpoint");
-        router = router.route(
-            "/api/v2/chat/completions-detection",
-            post(chat_completions_detection),
-        );
-    }
-
-    let app = router.with_state(state).layer(
-        TraceLayer::new_for_http()
-            .make_span_with(utils::trace::incoming_request_span)
-            .on_request(utils::trace::on_incoming_request)
-            .on_response(utils::trace::on_outgoing_response)
-            .on_eos(utils::trace::on_outgoing_eos),
-    );
-
-    // (2c) Generate main guardrails server handle based on whether TLS is needed
-    let listener: TcpListener = TcpListener::bind(&http_addr)
-        .await
-        .unwrap_or_else(|_| panic!("failed to bind to {http_addr}"));
-    let guardrails_handle = if let Some(tls_config) = tls_config {
-        // TLS
-        // Use more low level server configuration than axum for configurability
-        // Ref. https://github.com/tokio-rs/axum/blob/main/examples/low-level-rustls/src/main.rs
-        info!("HTTPS server started on port {}", http_addr.port());
-        let tls_acceptor = TlsAcceptor::from(tls_config);
-        tokio::spawn(async move {
-            let graceful = hyper_util::server::graceful::GracefulShutdown::new();
-            let builder = hyper_util::server::conn::auto::Builder::new(TokioExecutor::new());
-            let mut signal = std::pin::pin!(shutdown_signal());
-            loop {
-                let tower_service = app.clone();
-                let tls_acceptor = tls_acceptor.clone();
-
-                // Wait for new tcp connection
-                let (cnx, addr) = tokio::select! {
-                    res = listener.accept() => {
-                        match res {
-                            Ok(res) => res,
-                            Err(err) => {
-                                error!("error accepting tcp connection: {err}");
-                                continue;
-                            }
-                        }
-                    }
-                    _ = &mut signal => {
-                        debug!("graceful shutdown signal received");
-                        break;
-                    }
-                };
-
-                // Wait for tls handshake
-                let stream = tokio::select! {
-                    res = tls_acceptor.accept(cnx) => {
-                        match res {
-                            Ok(stream) => stream,
-                            Err(err) => {
-                                error!("error accepting connection on handshake: {err}");
-                                continue;
-                            }
-                        }
-                    }
-                    _ = &mut signal => {
-                        debug!("graceful shutdown signal received");
-                        break;
-                    }
-                };
-
-                // `TokioIo` converts between Hyper's own `AsyncRead` and `AsyncWrite` traits
-                let stream = TokioIo::new(stream);
-
-                let hyper_service =
-                    hyper::service::service_fn(move |request: Request<Incoming>| {
-                        // Clone necessary since hyper's `Service` uses `&self` whereas
-                        // tower's `Service` requires `&mut self`
-                        tower_service.clone().call(request)
-                    });
-                let conn = builder.serve_connection_with_upgrades(stream, hyper_service);
-                let fut = graceful.watch(conn.into_owned());
-                tokio::spawn(async move {
-                    if let Err(err) = fut.await {
-                        warn!("error serving connection from {}: {}", addr, err);
-                    }
-                });
-            }
-
-            tokio::select! {
-                () = graceful.shutdown() => {
-                    debug!("Gracefully shutdown!");
-                },
-                () = tokio::time::sleep(std::time::Duration::from_secs(10)) => {
-                    debug!("Waited 10 seconds for graceful shutdown, aborting...");
-                }
-            }
-        })
-    } else {
-        // Non-TLS
-        // Keep simple axum serve call for http version
-        let http_server = axum::serve(listener, app.into_make_service())
-            .with_graceful_shutdown(shutdown_signal());
-        info!("HTTP server started on port {}", http_addr.port());
-        tokio::task::spawn(async { http_server.await.expect("HTTP server crashed!") })
-    };
-
-    // (3) Launch each server as a separate task
-    let (health_res, guardrails_res) = tokio::join!(health_handle, guardrails_handle);
-    health_res.unwrap();
-    guardrails_res.unwrap();
-    info!("Shutdown complete for servers");
-    Ok(())
-}
-
-/// Configures and runs standalone health server.
-async fn run_health_server(
-    addr: SocketAddr,
-    state: Arc<ServerState>,
-) -> tokio::task::JoinHandle<()> {
-    info!("starting health server on {addr}");
-    let app = Router::new()
-        .route("/health", get(health))
-        .route("/info", get(info))
-        .with_state(state);
-    let listener = TcpListener::bind(&addr)
-        .await
-        .unwrap_or_else(|_| panic!("starting health server failed: could not bind to {addr}"));
-    let server =
-        axum::serve(listener, app.into_make_service()).with_graceful_shutdown(shutdown_signal());
-    tokio::task::spawn(async { server.await.expect("health server crashed!") })
-}
-
-async fn health() -> Result<impl IntoResponse, ()> {
-    // NOTE: we are only adding the package information in the `health` endpoint to have this endpoint
-    // provide a non empty 200 response. If we need to add more information regarding dependencies version
-    // or such things, then we will add another `/info` endpoint accordingly. And those info
-    // should not be added in `health` endpoint`
-    let info_object = HashMap::from([(PACKAGE_NAME, PACKAGE_VERSION)]);
-    Ok(Json(info_object).into_response())
-}
-
-async fn info(
-    State(state): State<Arc<ServerState>>,
-    Query(params): Query<InfoParams>,
-) -> Result<Json<InfoResponse>, Error> {
-    let services = state.orchestrator.client_health(params.probe).await;
-    Ok(Json(InfoResponse { services }))
-}
-
-async fn classification_with_gen(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<Json<models::GuardrailsHttpRequest>, Error>,
-) -> Result<impl IntoResponse, Error> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    request.validate()?;
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = ClassificationWithGenTask::new(trace_id, request, headers);
-    match state.orchestrator.handle(task).await {
-        Ok(response) => Ok(Json(response).into_response()),
-        Err(error) => Err(error.into()),
-    }
-}
-
-async fn generation_with_detection(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<
-        Json<models::GenerationWithDetectionHttpRequest>,
-        Error,
-    >,
-) -> Result<impl IntoResponse, Error> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    request.validate()?;
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = GenerationWithDetectionTask::new(trace_id, request, headers);
-    match state.orchestrator.handle(task).await {
-        Ok(response) => Ok(Json(response).into_response()),
-        Err(error) => Err(error.into()),
-    }
-}
-
-async fn stream_classification_with_gen(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<Json<models::GuardrailsHttpRequest>, Error>,
-) -> Sse<impl Stream<Item = Result<Event, Infallible>>> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    if let Err(error) = request.validate() {
-        // Request validation failed, return stream with single error SSE event
-        let error: Error = error.into();
-        return Sse::new(
-            stream::iter([Ok(Event::default()
-                .event("error")
-                .json_data(error.to_json())
-                .unwrap())])
-            .boxed(),
-        );
-    }
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = StreamingClassificationWithGenTask::new(trace_id, request, headers);
-    let response_stream = state.orchestrator.handle(task).await.unwrap();
-    // Convert response stream to a stream of SSE events
-    let event_stream = response_stream
-        .map(|message| match message {
-            Ok(response) => Ok(Event::default()
-                //.event("message") NOTE: per spec, should not be included for data-only message events
-                .json_data(response)
-                .unwrap()),
-            Err(error) => {
-                let error: Error = error.into();
-                Ok(Event::default()
-                    .event("error")
-                    .json_data(error.to_json())
-                    .unwrap())
-            }
-        })
-        .boxed();
-    Sse::new(event_stream).keep_alive(KeepAlive::default())
-}
-
-async fn stream_content_detection(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    json_lines: JsonLines<StreamingContentDetectionRequest>,
-) -> Result<impl IntoResponse, Error> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    // Validate the content-type from the header and ensure it is application/x-ndjson
-    // If it's not, return a UnsupportedContentType error with the appropriate message
-    let content_type = headers
-        .get(http::header::CONTENT_TYPE)
-        .and_then(|value| value.to_str().ok());
-    match content_type {
-        Some(content_type) if content_type.starts_with("application/x-ndjson") => (),
-        _ => {
-            return Err(Error::UnsupportedContentType(
-                "expected application/x-ndjson".into(),
-            ));
-        }
-    };
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-
-    // Create input stream
-    let input_stream = json_lines
-        .map(|result| match result {
-            Ok(message) => {
-                message.validate()?;
-                Ok(message)
-            }
-            Err(error) => Err(orchestrator::errors::Error::Validation(error.to_string())),
-        })
-        .enumerate()
-        .boxed();
-
-    // Create task and submit to handler
-    let task = StreamingContentDetectionTask::new(trace_id, headers, input_stream);
-    let mut response_stream = state.orchestrator.handle(task).await?;
-
-    // Create output stream
-    // This stream returns ND-JSON formatted messages to the client
-    // StreamingContentDetectionResponse / server::Error
-    let (output_tx, output_rx) = mpsc::channel::<Result<String, Infallible>>(128);
-    let output_stream = ReceiverStream::new(output_rx);
-
-    // Spawn task to consume response stream (typed) and send to output stream (json)
-    tokio::spawn(async move {
-        while let Some(result) = response_stream.next().await {
-            match result {
-                Ok(msg) => {
-                    let msg = utils::json::to_nd_string(&msg).unwrap();
-                    let _ = output_tx.send(Ok(msg)).await;
-                }
-                Err(error) => {
-                    // Convert orchestrator::Error to server::Error
-                    let error: Error = error.into();
-                    // server::Error doesn't impl Serialize, so we use to_json()
-                    let error_msg = utils::json::to_nd_string(&error.to_json()).unwrap();
-                    let _ = output_tx.send(Ok(error_msg)).await;
-                }
-            }
-        }
-    });
-
-    Ok(Response::new(axum::body::Body::from_stream(output_stream)))
-}
-
-async fn detection_content(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<
-        Json<models::TextContentDetectionHttpRequest>,
-        Error,
-    >,
-) -> Result<impl IntoResponse, Error> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    request.validate()?;
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = TextContentDetectionTask::new(trace_id, request, headers);
-    match state.orchestrator.handle(task).await {
-        Ok(response) => Ok(Json(response).into_response()),
-        Err(error) => Err(error.into()),
-    }
-}
-
-async fn detect_context_documents(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<Json<models::ContextDocsHttpRequest>, Error>,
-) -> Result<impl IntoResponse, Error> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    request.validate()?;
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = ContextDocsDetectionTask::new(trace_id, request, headers);
-    match state.orchestrator.handle(task).await {
-        Ok(response) => Ok(Json(response).into_response()),
-        Err(error) => Err(error.into()),
-    }
-}
-
-async fn detect_chat(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<Json<models::ChatDetectionHttpRequest>, Error>,
-) -> Result<impl IntoResponse, Error> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    request.validate_for_text()?;
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = ChatDetectionTask::new(trace_id, request, headers);
-    match state.orchestrator.handle(task).await {
-        Ok(response) => Ok(Json(response).into_response()),
-        Err(error) => Err(error.into()),
-    }
-}
-
-async fn detect_generated(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<
-        Json<models::DetectionOnGeneratedHttpRequest>,
-        Error,
-    >,
-) -> Result<impl IntoResponse, Error> {
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    request.validate()?;
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = DetectionOnGenerationTask::new(trace_id, request, headers);
-    match state.orchestrator.handle(task).await {
-        Ok(response) => Ok(Json(response).into_response()),
-        Err(error) => Err(error.into()),
-    }
-}
-
-async fn chat_completions_detection(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    WithRejection(Json(request), _): WithRejection<Json<ChatCompletionsRequest>, Error>,
-) -> Result<impl IntoResponse, Error> {
-    use ChatCompletionsResponse::*;
-    let trace_id = Span::current().context().span().span_context().trace_id();
-    let headers = filter_headers(&state.orchestrator.config().passthrough_headers, headers);
-    let task = ChatCompletionsDetectionTask::new(trace_id, request, headers);
-    match state.orchestrator.handle(task).await {
-        Ok(response) => match response {
-            Unary(response) => Ok(Json(response).into_response()),
-            Streaming(response_rx) => {
-                let response_stream = ReceiverStream::new(response_rx);
-                // Convert response stream to a stream of SSE events
-                let event_stream: BoxStream<Result<Event, Infallible>> = response_stream
-                    .map(|message| match message {
-                        Ok(Some(chunk)) => Ok(Event::default().json_data(chunk).unwrap()),
-                        Ok(None) => {
-                            // The stream completed, send [DONE] message
-                            Ok(Event::default().data("[DONE]"))
-                        }
-                        Err(error) => {
-                            let error: Error = error.into();
-                            Ok(Event::default()
-                                .event("error")
-                                .json_data(error.to_json())
-                                .unwrap())
-                        }
-                    })
-                    .boxed();
-                let sse = Sse::new(event_stream).keep_alive(KeepAlive::default());
-                Ok(sse.into_response())
-            }
-        },
-        Err(error) => Err(error.into()),
-    }
-}
-
-/// Shutdown signal handler
-async fn shutdown_signal() {
-    let ctrl_c = async {
-        signal::ctrl_c()
-            .await
-            .expect("failed to install Ctrl+C handler");
-    };
-
-    #[cfg(unix)]
-    let terminate = async {
-        signal::unix::signal(signal::unix::SignalKind::terminate())
-            .expect("failed to install signal handler")
-            .recv()
-            .await;
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        _ = ctrl_c => {},
-        _ = terminate => {},
-    }
-
-    info!("signal received, starting graceful shutdown");
-}
-
-/// High-level errors to return to clients.
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error("{0}")]
-    Validation(String),
-    #[error("{0}")]
-    NotFound(String),
-    #[error("{0}")]
-    ServiceUnavailable(String),
-    #[error("unexpected error occurred while processing request")]
-    Unexpected,
-    #[error(transparent)]
-    JsonExtractorRejection(#[from] JsonRejection),
-    #[error("{0}")]
-    JsonError(String),
-    #[error("unsupported content type: {0}")]
-    UnsupportedContentType(String),
-}
-
-impl From<orchestrator::Error> for Error {
-    fn from(value: orchestrator::Error) -> Self {
-        use orchestrator::Error::*;
-        match value {
-            DetectorNotFound(_) | ChunkerNotFound(_) => Self::NotFound(value.to_string()),
-            DetectorRequestFailed { ref error, .. }
-            | ChunkerRequestFailed { ref error, .. }
-            | GenerateRequestFailed { ref error, .. }
-            | ChatCompletionRequestFailed { ref error, .. }
-            | TokenizeRequestFailed { ref error, .. } => match error.status_code() {
-                StatusCode::BAD_REQUEST | StatusCode::UNPROCESSABLE_ENTITY => {
-                    Self::Validation(value.to_string())
-                }
-                StatusCode::NOT_FOUND => Self::NotFound(value.to_string()),
-                StatusCode::SERVICE_UNAVAILABLE => Self::ServiceUnavailable(value.to_string()),
-                _ => Self::Unexpected,
-            },
-            JsonError(message) => Self::JsonError(message),
-            Validation(message) => Self::Validation(message),
-            _ => Self::Unexpected,
-        }
-    }
-}
-
-impl Error {
-    pub fn to_json(self) -> serde_json::Value {
-        use Error::*;
-        let (code, message) = match self {
-            Validation(_) => (StatusCode::UNPROCESSABLE_ENTITY, self.to_string()),
-            NotFound(_) => (StatusCode::NOT_FOUND, self.to_string()),
-            ServiceUnavailable(_) => (StatusCode::SERVICE_UNAVAILABLE, self.to_string()),
-            UnsupportedContentType(_) => (StatusCode::UNSUPPORTED_MEDIA_TYPE, self.to_string()),
-            Unexpected => (StatusCode::INTERNAL_SERVER_ERROR, self.to_string()),
-            JsonExtractorRejection(json_rejection) => match json_rejection {
-                JsonRejection::JsonDataError(e) => {
-                    // Get lower-level serde error message
-                    let message = e.source().map(|e| e.to_string()).unwrap_or_default();
-                    (e.status(), message)
-                }
-                _ => (json_rejection.status(), json_rejection.body_text()),
-            },
-            JsonError(_) => (StatusCode::UNPROCESSABLE_ENTITY, self.to_string()),
-        };
-        serde_json::json!({
-            "code": code.as_u16(),
-            "details": message,
-        })
-    }
-}
-
-impl IntoResponse for Error {
-    fn into_response(self) -> Response {
-        use Error::*;
-        let (code, message) = match self {
-            Validation(_) => (StatusCode::UNPROCESSABLE_ENTITY, self.to_string()),
-            NotFound(_) => (StatusCode::NOT_FOUND, self.to_string()),
-            ServiceUnavailable(_) => (StatusCode::SERVICE_UNAVAILABLE, self.to_string()),
-            UnsupportedContentType(_) => (StatusCode::UNSUPPORTED_MEDIA_TYPE, self.to_string()),
-            Unexpected => (StatusCode::INTERNAL_SERVER_ERROR, self.to_string()),
-            JsonExtractorRejection(json_rejection) => match json_rejection {
-                JsonRejection::JsonDataError(e) => {
-                    // Get lower-level serde error message
-                    let message = e.source().map(|e| e.to_string()).unwrap_or_default();
-                    (e.status(), message)
-                }
-                _ => (json_rejection.status(), json_rejection.body_text()),
-            },
-            JsonError(_) => (StatusCode::UNPROCESSABLE_ENTITY, self.to_string()),
-        };
-        let error = serde_json::json!({
-            "code": code.as_u16(),
-            "details": message,
-        });
-        (code, Json(error)).into_response()
-    }
-}
-
-impl From<models::ValidationError> for Error {
-    fn from(value: models::ValidationError) -> Self {
-        Self::Validation(value.to_string())
-    }
-}
-
-fn filter_headers(passthrough_headers: &HashSet<String>, headers: HeaderMap) -> HeaderMap {
-    headers
-        .iter()
-        .filter(|(name, _)| passthrough_headers.contains(&name.as_str().to_lowercase()))
-        .map(|(name, value)| (name.clone(), value.clone()))
-        .collect()
 }
